test runner: deprecate `return_value`; switch to `response_value`
Response value better describes what is being returned.  Yes, it
is a return value of some sorts, but what's return value.  It could
be from the run method or it could be from a halt in the init method.

I proposing it is better to describe it more specifically as a
value that represents the type of response that would result from
this action.  Thus, `response_value`.

This deprecates the `return_value` method with a warning so that
pre-existing test code can be migrated to this new method.