url: take over prev Cgi functions; alter escaping behavior
In general, Deas shouldn't be forcing blanket query param escaping
on generated urls.  This makes query param escaping configurable
via the router.  This also simplifies the query string building
logic.  Deas no longer does special handling for array and hash
query param values.  All of this is to make Deas query param
handling customizable if desired.

The router now has an `escape_query_value` method that allows
defining a custom proc for escaping query values.  By default, rack
is used to escape values. The router has also been updated to allow
specifying a custom escape query value proc when adding a url. This
can be passed as an option when adding a url and it will be used
to escape the urls query values.

As mentioned, url generation no longer provide special handling
for arrays or hashes.  Previously, these would automatically be
flattened into a single hash with specially formatted keys.  The
goal of this was to generate urls that rack would convert into
arrays and hashes when it received a request using the url.  This
behavior is no longer automatically done.  This is to simplify
the url generation and avoid custom logic that doesn't make sense
for Deas.

This removes the Cgi functions as they are only used by the url
logic.  Also, the escaping isn't really Cgi-related anymore so it
shouldn't be named that way.  The remaining logic has been moved
into the url class where it is being used.