have redirects honor any router base url
This was missed when I originally added the redirect base url logic
back in PR 172.  In that effort I took care of honoring the base
url on the incoming path, but not on the redirect path.

This logic adds handling for the redirect path side of things.  To
accomplish this, redirect proxies need to be passed a router so
they can prepend the router's base url when generating their redir
paths.

In addition, I updated the router tests to randomly have a base url
set or not and to actually test the redirect paths that were being
produced.  The orig tests didn't do these things which contributed
to this logic being missed in the first place.  This involved adding
a `set_base_url` method allows manually resetting the base url to
`nil` since this can't be done wth the `base_url` method.