remove Sinatra from the error handling process
This updates the error handler to not depend on Sinatra.  This is
part of prepping to remove Sinatra from Deas.  This also gets Deas'
error handling more inline with how Sanford handles errors.  Ther
are three big differences this introduces.

First, error procs are no longer evaluated in Sinatra's scope.
Practically this means you can no longer do Sinatra-y things like
halting in your error procs.  Proc return values are still used
to alter responses but you'll have to treat these handlers more
like raw Rack and manually return the appropriate response values.
This also means that *all* error procs will always be run and the
last non-nil return value will always be used as the response.
Before, you could theoretically halt in an error proc and cutoff
any remaining error procs.

Second, error procs are now called with some context data in addition
to the exception that is being raised.  This data includes the
server data, the request/response and the handler class (if any).
Previously some of this data was implicitly available to handlers
from the fact that the proc was run in sinatra's scope.  This
switches to making certain pieces of data *explicitly* available.

Finally, this makes the *actual runner's* handler and params
available via the context data.  Previously, these attrs were not
available to error procs b/c you were in the scope of sinatra who
was not aware of handlers or the params given to the handler.  The
only difference between the handler params and the request params
is that the handler params are formally normalized by the runner.
In non-test cases they should be equivalent.  In general, you
should prefer the context params but fall back to the context
request params if there are no context params (ie the error
happened before any handler/runner was invoked).