Update summary logging to not error without a handler class
This updates the summary logging middleware to not error when it
is called and the env doesn't have a handler class. This happens
when an unknown route is called. Sinatra raises a not found error
and no handler class is set in the env.

This fixes the error by checking if the env key is set before
adding it to the line attrs. If there is no handler class, the
summary line will not contain a `handler`.

This also reworks the logging tests to fully test this logic.
Previously, we skipped fully testing our middleware because of the
setup required and because we were hurrying deas development.
Because of this, when the error was introduced the test suite
didn't error. This updates the tests to test more of the logging
middlewares logic and should help to avoid this happening in the
future.

As part of rewriting the tests, the fake sinatra call was updated
to have a `call` method so it can be used when testing the
middlewares. Its `call` method simply returns its response. The
response it builds was also updated to have values for its status,
headers and body. This avoids `nil` errors when those are used by
middlewares.

Finally, this also updates the test server used in the system tests
to randomly choose between summary and verbose logging. This will
also ensure that we randomly run both logging middleware and help
to catch errors that occur in them.