Cleanups for show exceptions and its tests
This is a small set of cleanups to the show exceptions middleware
and its tests to match our latest conventions. This is part of
cleaning up our middlewares and testing them consistently.

This simplifies the `Body` helper. It doesn't need to lazily set
its content, size and mime type because we use all of them when we
build it. This also reorganizes the tests to match our latest
conventions and switches to using the fake sinatra call for the
middlewares app. This also adds a test to make sure the apps
response isn't modified if there isn't an error.