router: lazy apply routes definitions
This switches the router logic to lazily apply route definitions.
The goal here is to allow defining a route that uses named urls
without first having defined the named url.  Overall, the router
should not be picky about the order directives are specified as long
as everything is valid once all directives have been specified.

To facilitate this, the route now only stores the definitions for
routes when a route or redirect method is called.  These definitions
build up and can be converted to actual routes on the router by
calling the `apply_definitions!` method.  The server now auto applies
definitions during its validation.

Note: to simplify the server and require it to know as little about
the router as possible, I chose to add a `validate!` method to the
router that the server now calls.  This method not only applies
any pending definitions, it also validates each one of its routes
like the server was doing previously.