remove usage of `env['sinatra.error']` and `Sinatra::NotFound`
This is another small piece in prep for removing Sinatra as a
dependency.  This does a few things:

* replaces `Sinatra::NotFound` exception handling with `Deas::NotFound`
* includes the path info in the not found exception message
* makes the exception factory randomly generate with `nil` backtraces
  as we've seen this come up in production
* make the error handling tolerate exceptions with `nil` backtraces
  gracefully

Overall, this is about removing sinatra usage.  In addition we get
the benefit of not found exceptions that actually include the path
that wasn't found (which is really helpful when debugging or when
these come up in production).