remove `response` as a handler/runner attribute
This effectively means you can't do direct rack response manipulation
in the handlers.  You must use a helper (ie status, headers, body, halt,
etc).

This is prep for adding a `response` handler helper for setting
responses (similar to `halt` except it won't halt execution).  This
helper will be how response data is set and will deprecate the
run-return-value-gets-set-as-the-response behavior.