cleanups to the system tests
This cleans up the system tests to get them up to our latest standards
and to get them all running/passing as intended.  The biggest change
is certain tests weren't returning a 200 status but we didn't assert
this and the other thing being tested just *happened* to pass but
wasn't passing b/c of the thing under test.

This is prep for removing sinatra from the response handling.  Sinatra
is the cause of the test *happening* to pass and this was exposed
when Sinatra was removed from response handling.