prep view handler for removing Sinatra from response handling
These reorganizes the private helper methods to group them by
their function.  This also updates the tests to get them up to our
conventions and to actually unit test the private helpers.  This
adds a bunch of tests that should have been in place originally but
were missed.

This was noticed while removing Sinatra from response handling.  This
just adds "noise" to that future effort so I'm breaking it off in
this PR to keep that coming effort as focused as possible.