view handler: namespace non-user-facing methods
Since view handlers are intended to run user code, we want to be
careful to not "pollute" the handler scope with our own methods
that will break in spectacular ways if the user happens to override
them.

This switches all the internal methods (ie methods not intended to
be called or used by users) to be prefixed with `deas_`.  This not
only marks them as internal "deas" methods but also makes it more
difficult to accidentally override in custom user handlers.

In addition, now that the methods are namespaced, I have no problem
making the `deas_run_callback` method public.  This allows the
deas runner to call the method directly instead of instance eval
calling it (which feels unnecessarily hackish).