hotfix - runners: cleanup incoming args hash handling
This does a couple of really minor cleanups to how the runners
handle their incoming args hashes.  The deas runner now does the
same handling but avoids creating unnecessary local variables. The
test runner switches to the `a` local variable for its dup'd args.
There is no technicaly reason to do this - this just avoids confusion
with the incoming args local var and highlights that the test runner
is altering the incoming args.

/cc @jcredding