cleanups: more thorough testing; consistency with Sanford and Qs
This touches up a bunch of things mostly related to testing.  The
goal here is to get Deas' implementation and consistant as possible
to Sanford and Qs.  All 3 of these gems implement the handler pattern
(ie their handlers and the runners that run them and the callbacks
they support) and should be quite similar in how the pattern is
implemented and tested.

I was updating Sanford and Qs making some of the recent updates
I've made to Deas and I noticed the discrepencies.  The biggest
change here is in the runner tests.  They now test with multiple
of each callback and test they were called in the expected order.

Everything else is super minor and is just obvious cleanups or style
fixes.