hotfix: relax rack/sinatra version requirements
The goal is to make Deas as flexible as possible to use with legacy
code bases.  In general the most modern versions should be used.
However if a legacy code base needs a legacy rack/sinatra version
(for example Rails 2/3) and only needs a subset of Deas logic (for
example just the rendering logic), there isn't really a good reason
to block it.

/cc @jcredding