template source: pass ext as an option to registered engines
This updates template sources to pass the extension that was used
to register an engine to the engine as an option. The goal is to
allow the engine to choose templates that match the extension that
was used when configuring the extension. This is also part of
fixing an issue with the engines not knowing which template file
to pick because they no longer hard-code an extension that the
templates have to use.

The template source now always sets an ext option and passes it
to the engines when they are registered. The ext will always be
set and can't be different from how the engine is registered. It
would be non-sensical to register with one engine ext and pass a
different engine ext as an option. Deas would either not use
the registered engine to render the template or the template
engine wouldn't find the file. Either case wouldn't produce the
desired result.

Specifically this helps address issues b/c Nm and deas-erubis no
longer expect specific file extensions but take optional file
extensions as arguments when looking up source files.  We want to
have the extension used to register the engine available to pass
in when they look up source files.