templates: complain if template name matches multiple files
This updates the template source and null template engine logic
to raise an argument error if the template name matches multiple
template files. This also changes how the template source globs
for template files when parsing the ext list. The goal is to make
Deas' template rendering more straightforward and avoid surprises
due to Deas choosing the first template it finds that matches.  Instead
this alerts the developer that they have requested a template using
a non-deterministic name and Deas isn't going to try and guess which
one they want.

By raising an error if multiple template files match, this avoids
any surprises with an unexpected file being rendered. This is
particularly nefarious if a development machines orders files one
way when globbing and a production machine orders them a different
way. In development everything would work as expected but then
fail in production.

Finally, this no longer disallows extensions for engine
registration. Previously, we would explicitly not allow rb engines
which would ensure that we didn't pick an rb file as a template.
This is really restrictive and doesn't really provide any
protection. If a user wants to use an template file with an rb
extension they should be allowed to.