templates: don't reject unregistered exts if they are the last ext on the file
Previously, we just blindly rejected all un-registered template exts.
This is nice b/c it means we don't incur the overhead of running
through the null engine repeatedly for the un-registered exts.

However, some engines require that source files end in their configured
ext when doing the initial render.  For example, say you have a template
named `template.erb.foo`.  The erb engine requires that its source
files end in `.erb`.  Without this change, `.foo` would get rejected
and the erb engine would be used to source render the file.  But the
engine would error b/c the file doesn't end in `.erb`.  None of this
makes sense to the user who just expects Deas to "ignore" unregistered
extensions.

With this change, the `.foo` ext in the example above is not rejected
which means the null template engine will source the file and pass
the contents off to the erb engine `compile` method (which imposes
no restrictions on the source file).  The erb engine compiles the
content and everything works as intended.

This change is the "best of both worlds": the edge-case scenario
above is handled as expected while useless unregistered exts like
`.html` in `template.html.erb` are ignored without having to run
through the null template engine compile method.