templates: cache null template engines on known null extensions
This minor optimization will now cache null template extensions.
The idea is to not repeatedly call the engine cache default block
for known null extensions.

To make this possible, I had to update the logic for determining
if we have a registered engine.  We now store a separate list of
registered extensions. This implementation is nice b/c it decouples
the engine cache from defining whether an engine is registered or not.

This will be helpful once PR 209 is merged as templates ending with
null extensions will won't have those extensions removed from their
extension list and would otherwise incur the overhead mentioned
above.