template source: minor optimization to private `compile` logic
The goal here is to create less objects and avoid having to modify
variable state.  This drops the `dup` and `shift` operations in
favor of more functional approaches using `drop(1)` and `first` to
get the necessary parts of the extension list.  I like this better
b/c it doesn't mutate state and doesn't require adding a local var.

The tests don't change and everything works as it did before.  There
doesn't seem to be any performance impact either.