remove ns-options dependency
We are moving away from using ns-options as a dependency b/c it
has performance issues when accessing options and on top of that
its API hasn't proven to be especially useful over a struct-like
class.

This removes ns-option and reworks the server, its config and the
sinatra app to support the non-ns-options config implementation.
The config is now a simple struct-like class with a bunch of
accessors for its attrs.  The server class methods now implement
the ns-options-like DSL api for each of their attrs for backwards
compatibility.

In addtion, I did a number of smaller changes/cleanups, including
some thing that aren't backwards compatible:

* switched to calling the config "config" (more succinct than prev
  with no loss of clarity).
* you can no longer set config values to procs (for lazy eval'ing).
  I chose to not keep this behavior.
* I had to initially set the `template_source` to `nil` and then
  overrides the reader method to default it if it was still `nil`
  when being read.  This hacks the lazy proc behavior from ns-options
  and allows the default template source to be based on whatever
  root is set at the time it is referenced.
* I changed the config/server api to set view/public *path snippets*
  instead of fully-qualified paths strings.  Any path values get
  expanded relative to the root setting.  I wanted to drop the
  pathname requirement and this allows you to set root relative
  snippets for these settings.
* I dropped the `to_hash` logic and switched to just manually
  building the server data in the sinatra app.  I'm just trying
  to simplify the API here and blindly using the config's hash
  kinda hides what values are being set.
* I reworked the order of config/server/sinatra app settings to
  try and get everything to match.  This is just to be consistent
  and help keep things organized/readable.
* I did a bunch of test cleanups to not only test the new logic
  but use modern conventions like Factory values, etc.
* updated the assert dependency to get access to the latest stuff.