updated body handling to better adhere to Rack standard
This is prep not only for removing Sinatra as a dependency but also
fixes some bugs in modern ruby versions (ie non-1.8.7).  In 1.8.7,
`String` responds to `each` but in modern versions it does not.  This
change now forces the body values to always respond to `each` (per
Rack spec) by forcing string bodies to be transformed to arrays. As
a side-effect, the `:each` check and custom exception was removed.

This also does a number of minor tweaks related to this change and
also to get the test suite passing in modern ruby versions:

* update to the latest assert patch - this version has a minor fix
  for modern ruby versions
* `halt` now *assumes* any arg after the status and headers checks
  is the body value - no need to perform the `:each` check as it
  is unreliable for string bodies
* the runner-related tests now expect array body values (per Rack
  spec)
* the url tests now build with arrays for both `:captures` and
  `:splat` params.  This should have been done originally as this
  is Sinatra spec, but `String#each` in 1.8.7 was hiding this.