urls: have `url_for` complain if passed empty values for named params
Empty (ie `nil` or empty string) values aren't valid for named param
values as urls with named params are never matched on empty values
(the route just won't match at all).  Therefore it is silly to
build urls with empty values as they will never be matched.

This switches the `url_for` method on the Router to now complain
if given an empty named param value.  The goal here is to give the
developer some warning that they are building a non-sensical url
value.  Note: this doesn't apply to splat param values or any "extra"
param values (which are added to the query string).  This only
applies to formally named params in the url path.