server: call router methods by evaluating a block on the `router` method
This is an update to simplify the server's API and to get Deas'
server implementation more inline with Sanford's server and Qs'
daemon.  Using this implementation has a couple of advantages:
we don't have demeter all the router methods keeping them in sync
with the router and implicitly namespaces the router methods in
the server api.

This implementation matches what Sanford and Qs do.  Our goal is
to keep each of these "app interface" gems' implementation as
consistent as possible.

Note: the `url_for` method was kept in the server api so that users
wouldn't *have* to always call it via the server's router object.
This method is actually intended to be called as a helper function
in app code so we want to keep it as easy to use as possible.

Note: this also does a few cleanups to the router, router tests and
template source tests.  There are things I noticed while updating
Sanford's code using these as reference.