formalize restricted splat handling
This reworks splat handling in Deas.  Originally, Deas handled
splats just as Sinatra would.  However, the plan was never to
support the full splat flexibility of Sinatra's splats in Deas'
upcoming Sinatra-free implementation.

So, back in PR 181, we removed both the splat and any "captures"
from the params Sinatra provided.  The router (being Sinatra's)
still honored splat routes but you couldn't access the splat
values in the params.  This was Step 1 in getting away from the
Sinatra splat implementation.

However, there are undeniable uses for splats and it was a little
naive to assume that no splat handling is acceptable for Deas.
This effort is an attempt to take the best parts of Sinatra's
splat handling and provide that in Deas.  This comes in the form
of two features:

* limit splats to just a single splat at the end of a url path
* captures the splat value and makes it available to view handlers
  as a `splat` helper method

The first feature above is an attempt to simplify router
requirements by only allowing splat parsing on the end of url
paths.  The second feature allows not limiting the param names
that can be used while still providing the parsed splat value.

The end result is the public API for splat handling.  As we remove
Sinatra in future efforts, these are the features we need to
continue supporting.