remove Sinatra from route/request handling
This is prep for removing Sinatra as a dependency in the future.
This not only removes Sinatra from the request handling but also
decouples the rack app implementation from request handling by
introducing the concept of "request data" and passing that
between the rack endpoint and the runner/handler.

So, this creates a RequestData struct.  Right now the request data
consists of the request, response, params and route path.  The
route path is being added to not only have for logging/error usage
but to also use in a coming effort to take over splat handling.
I chose to remove manual session handling as you can access the
session info on the request and I don't want deas to support
manual session handling going forward.

Note: this removes `sessions` handling form the server api and
its config.  We seemed to not be explicitly testing this though
so no tests had to be changed.

Note: I chose to keep the session related system tests to prove
that removing session support from Deas doesn't mean you can't
use vanilla rack sessions in Deas apps.

Note: this updates to a more restrictive rack dependency.  I want
to limit this going forward to eliminate edge-cases but also b/c
this is the oldest version that works with Rack::App (which I
plan to replace Sinatra with in the future).