take over error handling fully from Sinatra
This is prep for removing Sinatra in a future effort.  This
switches to rescuing StandardError on route runs instead of
configuring a sinatra error handler.  This takes Sinatra out of
the error handling logic.  Sinatra had already been removed from
the error handler class itself so this change is fairly trivial.

There is one small side-effect of this change: only a set of
"standard error" exceptions are rescued (which is different than
Sinatra's default of all exceptions).  The intent is that any
errors you would handle should already be StandardErrors as non
standard errors include system interrupts etc which are dangerous
to rescue as this may alter system behavior.  We include a few
common non-standard errors that can be expected to occur.  The
system tests have been updated to account for this; otherwise
they remain unchanged and valid.