add the route path to the error handler context data
Since we are now passing this data to the runner/handler
(see PR 219), it would be nice to also make it available in the
error handler context data.  Some applications include getting
extra context on where params are coming from and knowing which
route caused the error (in the case where multiple route paths
route to the same handler).

Note: I reorder the route path in the request data and where it
is used to keep things more consistently ordered.