update router route error checking to be more thorough
Back in PR 217, Deas was updated to restrict splats in route path
definitions to only allow trailing splats.  In that effort the
check was applied to named urls that were defined but if the route
didn't use a named url, the check was missed.  This moves that
error check to where the routes are added.  This way, no matter
where the path comes from, the check will be performed.

This also adds a new check to ensure named urls are defined with
symbols.  Before, any given name would be turned into a symbol
so if you defined a named route with a string you'd have to know
to refer to it with a symbol.  This removes that magic knowledge
requirement and just warns you if you aren't using symbols when
defining named urls so there won't be any confusion when you go
to use that named url.