add splat value to middlewares and error handler context info
This makes the splat value (if any) available to error handlers
as contextual info.  Since splats are no longer part of the params
we need a way to access this info in error handlers.

This also updates the handler proxy to store the splat value on
the request env.  This makes it available to middlewares.  The
also updates the logging (includeing the Logging middleware)to use
the splat info in their logs.  I chose to also log the route path
in the verbose logs for completeness (this should have been done
in previous efforts).

Note: I updated the error context to "fetch" its attrs from the
given args hash.  This is to enforce that if the context expects
a value the rack app supplies it.  This also closes a gap in our
tests and tests that the rack app is creating the context as
necessary.