update url generation to work with composed param names
This fixes a bug where if you happen to define a url with param
names that compose another params' names, the url generation will
not set the param names correctly.  This updates the tests to
test this case and updates the implementation to apply longer
param names before shorter param names.

I discovered this problem while working on PR 224.