update route checks to only allow last path segment splats
Before we were just checking that a splat was the last character
in a path string.  This isn't a good enough check b/c you could
use `/something*` and it was pass the check.  For our upcoming
custom router work, we need to ensures splats are always the
final path segment of a route.

This updates the regex and tests.  This also switches to using a
custom exception class so it can be caught and reraised in the
`Server#new` method.  This abstracts the Deas internals from the
error backtrace.