(hotfix) remove charset for json handling
Deas now automatically adds charset to the content type on all
renders.  Plus, deas-json specifically handles this case if it
is used.  Both of these were added in previous efforts but I just
forgot to go in and remove this setting (it was doing no harm).
I noticed this while prepping Deas to remove Sinatra.