(hotfix) cleanup the error handler tests setup
This fixes an issue with setting up the error handler tests that
I introduced back in #185.  I was confusing the "handler class"
meaning the error context value with the "handler class" meaning
`Deas::ErrorHandler` used in the test setup.

This cleans up the mistake and uses a more specific var name for
the error handler class to keep the confusion from coming up
again.  There were no behavior issues or side-effects to this
change - just a cleanup.