Allow adding before/after route run procs
This allows adding before/after route run procs when configuring
a server. These are needed for logic that you want to run before
or after every route but not run for every request via a
middleware. These will also run even if there is an exception.
This is specifically useful for adding logic to check in
activerecord connections after a route has been run in a threaded
environment. This doesn't need to run for requests that are
terminated by a middleware that never uses activerecord but it
needs to run for every route which has a high chance of using
activerecord.

A `Route` will now call before and after route run procs when its
run. It gets the procs from its `ServerData` and uses an `ensure`
to run the after procs. This means the procs will run even if an
exception occurs while running the handler proxy. The route also
passes the server data and request data to the procs. The `Server`
was also updated to have DSL methods for adding before and after
route run procs. These are passed to its server data which is how
the route gets the configured procs.