(hotfix) revert back to a less restrictive rack dependency
The more aggressively restrictive dependency was added in 8e97ab4
with the goal of limiting edge-case issues and also in prep for
removing Sinatra.

However, this restriction keeps from using Deas in old Rails 2.X
and 3.X apps.  This removes the restriction in favor of using in
Rails since the restriction and how it affects getting off of
Sinatra is not known right now.