update the `body` helper to properly implement the Rack spec
See www.rubydoc.info/github/rack/rack/master/file/SPEC#The_Body.
"The Body must respond to each and must only yield String values".
This is a minor tweak to for any single values into an array
*of strings* (since the body must yield strings).  We still want
to support allowing specifying bodies with single objects from the
Deas API point-of-view.  This just makes sure we are properly
adhering to Rack's spec.

I added comments to the code to help explain this so we are aware
in the future.