cleanup middleware args handling
So previously I had said that there was no good way to get server
config args to the middlewares except to store those on the
Sinatra app since this "app" was what was initialized with each
middleware.  Well, turns out I was wrong.

First off, the "app" initialized with each middleware is really
more like the "next upstream middleware or the end app".  I get
why the convention is to call it "app" b/c from the middleware's
point of view it might as well be the app, but this had confused
me.  Second, there is an interface for passing args to middlewares.
Sinatra's `use` takes the same `Rack::Builder` interface which
means both `*args` and a proc can be passed.  They just need to
be specified as addition params.

So, this switches to using this interface and also switches to
directly passing in the server's logger to the `Logging`
middleware.

This is a cleanup and is prep for adding a new optional middleware
for handling trailing slashes based on router settings.  It will
have to have the router passed in so it will init similar to how
the `Logging` middleware is init here.