update the router api for setting trailing slashes handling
This adds 3 new router api methods: `require_trailing_slashes`,
`require_no_trailing_slashes`, and `allow_trailing_slashes`.
Calling any of these methods will set a "trailing slashes" config
setting value on the router that will direct it how to handle
trailing slashes in its routes.  The goal with this is to help
prevent unnecessary 404s b/c a user (or the browser) accidentally
adds/omits a trailing slash on a URL path.

First let's look at `require_trailing_slashes`.  This is intended
when you want all of your paths to end in a trailing slash.  The
router will ensure this and error if it finds routes not ending
in a slash when validating.  In a future effort, this directive
will also force-redirect any paths not ending in a slash to end
in a slash.

`require_no_trailing_slashes` does the opposite: the router will
complain if any of your defined routes have trailing slashes and
in a future effort the directive will force-redirect any paths
ending in a slash to not ending in a slash.

`allow_trailing_slashes` won't complain about the router's routes -
you can freely mix and match routes ending in slashes and not. In
a future effort, Deas will switch the slash on paths and retry
them if it sees any `Deas::NotFound` errors.  This will "allow"
specifying and using either trailing slashes or not.

Finally, if you don't use any of the directives, everything
behaves as normal.  You will get 404s for paths that don't
exactly match what is defined in the router.

As mentioned this is the first of a two-part effort.  This sets
up the directives and the router's API.  In the next effort,
I'll add a middleware that will handle enforcing any router
directive's behavior.