middleware for trailing slashes
This adds a middleware that enforces any trailing slashes
directive in the Deas app's router.

This middleware takes a router as an arg and validates that it has
a trailing slashes directive set.  On call, the middleware looks
up a handler for the router directive and runs it.  I chose to
lookup the handlers from a hash to keep the logic at constant
time performance and to avoid a bunch of conditionals.  By org'ing
the logic in handlers, each directive's logic can be decoupled from
the other directives.

The require handler checks for a slash at the end of the path and
if there isn't one, redirects to the same path with a slash. The
require no handler does the opposite.  The allow handler tries the
path as is.  If it sees a `Deas::NotFound` (router not found)
response, it clears the error and retries the request with the
ending slash state switched.

I chose redirects for the require handlers to actually change the
url in the user's browser and to direct the browser to the desired
path representation.  I chose to retry (and to not redirect) in the
allow handler so that the url wouldn't change in the browser and
it would get the desired behavior as efficiently as possible.

Finally, this adds the middleware, if a router directive has been
set, to the server's middleware stack.  If no directive has been
set, no special handling is needed.  It is added last so that the
details of the allow retries as hidden from the logging middleware
and so that any errors will get the show exceptions logic.

This is part 2 of the main development effort for this new
"trailing slashes" feature.  I actually have one more thing to
handle after this though: I need to account for paths that end
in file extensions and not run any trailing slash behavior on them
as it is implies that they won't end in a slash no matter what
any directive says.  I'm doing that part next.