(hotfix) remove 'require trailing slashes' router directive
After talking through the edge cases of this feature (especially
what to do with requiring trailing slashes on file extension routes
(ie `/some/file.jpg/`??), it became more evident that this was a
non-feature.  The edge cases became complex and I couldn't think
of a scenario where you would want to *enforce* that all urls had
a trailing slash.  I could see cases where you might want a mix
but in this case either just define your routes in the mix you
want or just use the "allow" directive.

So, this removes that directive and all logic in the router for
validating it and all logic in the middleware for enforcing it.