jux "Route: ..." verbose logging after "Path: ..." logs
This is a minor tweak I noticed while reviewing logs to make sure
the trailing slash redirects were working properly.  Jux'ing the
route log immediately after the path log makes it easier to see
how the path matches the route and produces param values for the
route placholder values.

In addition, I updated all places where the route path is used or
referenced to also order similar to how the logs read.  This works
out nice since the router path drives the handler class and params
so it (generally) should appear in the code before any handler or
params stuff.  This did add a bunch of diff noise to this effort
though.

Here is an example showing the before/after in the logs from an
app I'm working on (these logs also show the trailing slash
redirects in action):

Before:
```
[INFO ] : [Deas] ===== Received request =====
[INFO ] : [Deas]   Method:  "GET"
[INFO ] : [Deas]   Path:    "/people/"
[INFO ] : [Deas]   Redir:   /people
[INFO ] : [Deas] ===== Completed in 0.01ms (302, FOUND) =====
[INFO ] : [Deas] ===== Received request =====
[INFO ] : [Deas]   Method:  "GET"
[INFO ] : [Deas]   Path:    "/people"
[INFO ] : [Deas]   Handler: Deas::RedirectHandler
[INFO ] : [Deas]   Params:  {}
[INFO ] : [Deas]   Route:   "/people"
[INFO ] : [Deas]   Redir:   /people/user-browser?status=Active
[INFO ] : [Deas] ===== Completed in 3.52ms (302, FOUND) =====
```

After:
```
[INFO ] : [Deas] ===== Received request =====
[INFO ] : [Deas]   Method:  "GET"
[INFO ] : [Deas]   Path:    "/people/"
[INFO ] : [Deas]   Redir:   /people
[INFO ] : [Deas] ===== Completed in 0.04ms (302, FOUND) =====
[INFO ] : [Deas] ===== Received request =====
[INFO ] : [Deas]   Method:  "GET"
[INFO ] : [Deas]   Path:    "/people"
[INFO ] : [Deas]   Route:   "/people"
[INFO ] : [Deas]   Handler: Deas::RedirectHandler
[INFO ] : [Deas]   Params:  {}
[INFO ] : [Deas]   Redir:   /people/user-browser?status=Active
[INFO ] : [Deas] ===== Completed in 3.6ms (302, FOUND) =====
```